pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Alex <kudinov-ao@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.23.2"
fastapi = "^0.103.1"
asyncpg = "^0.28.0"
pytest = "7.4.0"
httpx = "^0.25.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"
alembic = "^1.12.0"
psycopg2 = "^2.9.8"
python-dotenv = "^1.0.0"
pydantic = "2.4.2"
pydantic-settings = "2.0.3"
sqlalchemy = "2.0.21"
awesome-module = "^0.1.0"
pydantic-core = "2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
